+++ src/main/java/org/songeun/petdongne_server/global/converter/TransactionTypeConverter.java
@@ -2,7 +2,7 @@ package org.songeun.petdongne_server.global.converter;
 
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
-import org.songeun.petdongne_server.domain.residentialComplex.entity.TransactionType;
+import org.songeun.petdongne_server.residentialComplex.entity.TransactionType;
 
 @Converter
 public class TransactionTypeConverter implements AttributeConverter<TransactionType, String> {
