+++ src/main/java/org/songeun/petdongne_server/review/entity/ReviewPhoto.java
@@ -1,8 +1,8 @@
-package org.songeun.petdongne_server.domain.survey.entity;
+package org.songeun.petdongne_server.review.entity;
+
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
@@ -10,24 +10,27 @@ import org.songeun.petdongne_server.global.common.BaseEntity;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class SurveyOption extends BaseEntity {
+public class ReviewPhoto extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
     private Long id;
 
     @NotNull
-    @Size(min = 1, max = 150)
-    private String optionText;
+    private String url;
+
+    @NotNull
+    private Boolean petPhoto;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "survey_question_id")
-    private SurveyQuestion question;
+    @JoinColumn(name = "pet_residence_review_id")
+    private PetResidenceReview review;
 
     @Builder
-    private SurveyOption(String optionText, SurveyQuestion question) {
-        this.optionText = optionText;
-        this.question = question;
+    private ReviewPhoto(String url, Boolean petPhoto, PetResidenceReview review) {
+        this.url = url;
+        this.petPhoto = petPhoto;
+        this.review = review;
     }
 
 }
