+++ src/main/java/org/songeun/petdongne_server/survey/entity/SurveyOption.java
@@ -1,4 +1,4 @@
-package org.songeun.petdongne_server.domain.survey.entity;
+package org.songeun.petdongne_server.survey.entity;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
@@ -10,18 +10,24 @@ import org.songeun.petdongne_server.global.common.BaseEntity;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class SurveyQuestion extends BaseEntity {
+public class SurveyOption extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
     private Long id;
 
     @NotNull
-    @Size(min = 1, max = 255)
-    private String questionText;
+    @Size(min = 1, max = 150)
+    private String optionText;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "survey_question_id")
+    private SurveyQuestion question;
 
     @Builder
-    private SurveyQuestion(String questionText) {
-        this.questionText = questionText;
+    private SurveyOption(String optionText, SurveyQuestion question) {
+        this.optionText = optionText;
+        this.question = question;
     }
+
 }
