+++ src/main/java/org/songeun/petdongne_server/survey/entity/SurveyQuestion.java
@@ -1,7 +1,8 @@
-package org.songeun.petdongne_server.domain.user.entity;
+package org.songeun.petdongne_server.survey.entity;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
@@ -9,18 +10,18 @@ import org.songeun.petdongne_server.global.common.BaseEntity;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ProfileImage extends BaseEntity {
+public class SurveyQuestion extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
     private Long id;
 
     @NotNull
-    private String url;
+    @Size(min = 1, max = 255)
+    private String questionText;
 
     @Builder
-    private ProfileImage(String url) {
-        this.url = url;
+    private SurveyQuestion(String questionText) {
+        this.questionText = questionText;
     }
-
 }
